eks/cdk/requirements.txt
aws-cdk.aws-autoscaling
aws-cdk.aws-cloudwatch
aws-cdk.aws_ec2
aws-cdk.aws_eks
aws-cdk.aws-iam
aws-cdk.aws-fis
aws-cdk.aws-ssm
aws-cdk.core
pyyaml
